requirements.txt
numpy
pandas
scipy
pillow
matplotlib
scikit-learn
tensorflow
pyarrow
transformers
torch --index-url https://download.pytorch.org/whl/cu126
torchvision --index-url https://download.pytorch.org/whl/cu126
torchaudio --index-url https://download.pytorch.org/whl/cu126
ruff